requirements.txt
python-telegram-bot>=20.0
httpx
psycopg2-binary
